+++ CHANGELOG.md
@@ -14,6 +14,10 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   - The Types section of the documentation has been split up into separate sections for lattice
 basis vectors, atoms and crystal representations, and data grids.
 
+### Fixed
+  - The default definition of `Electrum.DataSpace(x)` is now `DataSpace(typeof(x))`, not
+`DataSpace(typeof(basis(x)))`.
+
 ## [0.1.16]: 2023-11-16
 
 ### Added